search: Fix build on 32-bit systems
Match the type in the header and source file for
gb_search_display_group_set_count() to use guint64, not a mixture of
guint64 and gsize (which is a typedef to unsigned long, and 32 bits wide
on 32-bit systems).

https://bugzilla.gnome.org/show_bug.cgi?id=743402

+++ src/search/gb-search-display-group.h
@@ -52,7 +52,7 @@ void              gb_search_display_group_add_result    (GbSearchDisplayGroup *g
 void              gb_search_display_group_remove_result (GbSearchDisplayGroup *group,
                                                          GbSearchResult       *result);
 void              gb_search_display_group_set_count     (GbSearchDisplayGroup *group,
-                                                         gsize                 count);
+                                                         guint64               count);
 void              gb_search_display_group_unselect      (GbSearchDisplayGroup *group);
 void              gb_search_display_group_focus_first   (GbSearchDisplayGroup *group);
 void              gb_search_display_group_focus_last    (GbSearchDisplayGroup *group);
